+++ Rei/src/Ecs/World.cpp
@@ -8,7 +8,7 @@ namespace rei::ecs
     World::World(): _ecsRegistry(std::make_shared<EcsRegistry>()),
                     _filterRegistry(std::make_shared<FiltersRegistry>())
     {
-        _ecsRegistry->MaxComponentIdChangedEvent += std::make_shared<std::function<void(u32)>>([this](const u32 s) { UpdateBitMasks(s); });
+        _ecsRegistry->MaxComponentIdChangedEvent += std::make_shared<std::function<void(size_t)>>([this](const size_t s) { UpdateBitMasks(s); });
     }
 
     void World::Run()
@@ -52,7 +52,7 @@ namespace rei::ecs
         return _filterRegistry;
     }
 
-    void World::UpdateBitMasks(const u32 size) const
+    void World::UpdateBitMasks(const size_t size) const
     {
         if (size < sizeof(BitMask::mask) * 8) return;
